Namespace support: initial commit to support namespace and use clauses
The Tolerant PHP Parser has an api to get the import tables for current scope. However, this api only works for anything after the NamespaceUseDeclaration, which does not work for the goal of this project. Because to get the most out of the performance, the server will only traverse the AST once, hence a custom function to read the import tables is created even though the majority of the code is taken from the API function